requirements.txt
# Core Data Processing
numpy==1.26.4
pandas==2.2.0
pyarrow==15.0.0

# PySpark
pyspark==3.5.0
delta-spark==3.0.0

# ML Monitoring
scipy==1.12.0

# Database - PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data Simulation
Faker==22.6.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# AI Agents (Pydantic AI + RAG + API)
pydantic-ai>=0.0.40
openai>=1.12.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1

# LLM Observability
langfuse>=2.0.0

# LLM Evaluation
deepeval>=3.8.0

# Multi-Agent Orchestration
langgraph>=0.3.0
langgraph-checkpoint-postgres>=2.0.0

# Azure AI Services
azure-identity>=1.15.0
azure-search-documents>=11.4.0
azure-ai-ml>=1.12.0
azureml-mlflow>=1.55.0
